[BBC-9100] Migrate Textarea to new Storybook
