+++ src/textarea/Textarea.unit.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { mount, shallow } from 'enzyme'
 
 import { EyeIcon } from '../icon/eyeIcon'
-import { Textarea } from './Textarea'
+import { Textarea } from './index'
 
 it('Should have the proper value.', () => {
   const wrapper = shallow(<Textarea name="test" defaultValue="blabla" />)
